Create GetIdByCpeURI endpoint
This creates an GetIdByCpeURI endpoint on the DB interface.

The rdb implementation splits the existing `GetByCpeURI` call into an `GetIdByCpeURI ` and the existing `GetMulti`. This reduces the duplication, while exposing intermediary data.

(I’m not sure if this is desired, but I found it useful for the tools I’m writing)